fix recent json database change of content_type for .tar.gz files
GloriousEggroll changed for the recent versions the json description for `content_type` from `"application/gzip"` to `"application/octet-stream"` for the tar.gz files. This content type is the same for the checksum files and therefore the check would not work anymore. So I added a quick fix checking the filename of each asset too.