fix the last fix for content_type of checksum files too
In the last fix I talked about the upstream change of content_type to `"application/octet-stream"` for recent database information. However it is actually changed to `"binary/octet-stream"`. This resulted into ignoring checksum_filename, as it still checked for the old content_type only and prevented the installation of the newest GE-Proton.

Therefore this fix is an extension of the last fix to make it complete.