refresh cached .summary for entries that are not ready
The downloadable archives to install from GE-Proton source are uploaded to Github at later time, even if the entry for releases is available already.

Which is a problem for my script, because it creates cache files for quick access and would not notice the change on the server. Therefore it needs to remove those *.summary files in cache folder, which have a 0 size for downloads. It's enough if this is done once at start of script.